Added a Map method to the task pool
Added a Map method to the task pool. This allows to give a vector of
values, together with a lambda function, and return a vector of the
results of the lambda. The order of the items will be preserved but
the lambdas are executed in parallel.

Example:
``` c++
Construction::Common::TaskPool pool (64); // Task pool running on 64
threads

std::vector<int> input = { 1, 2, 5 };

std::vector<int> result = pool.Map<int, int>(input, [](const int& value)
{ return value * value; });
// => { 1, 4, 25 }
```