Tensor contractions
* Replaced assert with exception for index access.

* Again, removed some asserts.

* Added a Map method to the task pool

Added a Map method to the task pool. This allows to give a vector of
values, together with a lambda function, and return a vector of the
results of the lambda. The order of the items will be preserved but
the lambdas are executed in parallel.

Example:
``` c++
Construction::Common::TaskPool pool (64); // Task pool running on 64
threads

std::vector<int> input = { 1, 2, 5 };

std::vector<int> result = pool.Map<int, int>(input, [](const int& value)
{ return value * value; });
// => { 1, 4, 25 }
```

* Replaced the IsEqual(...) method in Symmetrize with more intelligent heuristics

* Directly write the values into the components matrix in Simplify. This makes RAM use more efficient

* Fixed the bottleneck in Simplify

The mutex was locked too often in the insertion of the matrix
components. We only need to lock if the value is calculated and the
result is != zero. This will improve the performance tremendously.

* Added *.cs and *.dat to gitignore

Added some additional lines to gitignore

- ignore *.cs files for construct sheets

- ignore *.dat files for sessions

- ignore everything by Visual Studio Code

* Make the progress bar work with atomics such that the mutex does not need

* Added contravariant printing of tensors

* Added Kronecker delta tensor

* Removed the linking in Makefile since it wasn't working anyway

* Improved Simplify

* Improved printing for indices

* Stop reading off from the matrix if one row is completely zero

* Fixed issue #4.

* Remove zero values from the matrix, also in the Gauss-Jordan algorithm

* Improved Simplify

* Ignore .csd files

* Fixed the output bug in Kronecker delta

* Implemented tensor contraction.