Fixed bug in the apple program
There was a bug in the coefficient parser finite state machine and thus
an equation with more than one coefficient always added to reference to
the same coefficient. This lead to the solving method being invoked more
than once and thus the process was killed.

Furthermore the results of one equation will now be included
automatically in the next equation.