Fixed linux bug
* Added default constructor to CLI

* Added source file for the Apple program that allows solving a given list of equations to find the common solutions

* Added method to extract all the variables from a scalar expression

* Made tensor substitution better

* Implemented SetIndices for multiplications

* Added printing of the coefficients

* Added clean method to the progress bar and joining of the printing thread to avoid error messages at shutdown.

* Finished equation solving

* Fixed bug in the apple program

There was a bug in the coefficient parser finite state machine and thus
an equation with more than one coefficient always added to reference to
the same coefficient. This lead to the solving method being invoked more
than once and thus the process was killed.

Furthermore the results of one equation will now be included
automatically in the next equation.

* Fixed failing compilation on linux

Fixed failed compilation on linux. memory and algorithm were missing
in coefficients.hpp, equations.hpp and scalar.hpp